// src/interfaces/apiInterface.tsx
// Generated by https://quicktype.io

export interface APIList {
  status:  string;
  message: string;
  data:    Data;
}

export interface Data {
  id:           number;
  ipData:       IPData;
  isFavourited: boolean;
  isBanned:     boolean;
  isDeleted:    boolean;
  updatedAt:    string;
  createdAt:    string;
  conversion: string
}

export interface IPData {
  ipAddress:   string;
  isoCode:     string;
  countryName: string;
  regionName:  string;
  cityName:    string;
  latitude:    number;
  longitude:   number;
  countryFlag: string;
  currency:    string;
  conversion?: string
}